🚚 move: 패키지 상세 분리
